chore: refactor for maintainability/extendability
