+++ includes/class-newspack.php
@@ -80,6 +80,10 @@ final class Newspack {
 		include_once NEWSPACK_ABSPATH . 'includes/oauth/class-google-oauth.php';
 		include_once NEWSPACK_ABSPATH . 'includes/oauth/class-google-services-connection.php';
 
+		include_once NEWSPACK_ABSPATH . 'includes/starter_content/class-starter-content-provider.php';
+		include_once NEWSPACK_ABSPATH . 'includes/starter_content/class-starter-content-generated.php';
+		include_once NEWSPACK_ABSPATH . 'includes/starter_content/class-starter-content-wordpress.php';
+
 		include_once NEWSPACK_ABSPATH . 'includes/wizards/class-setup-wizard.php';
 		include_once NEWSPACK_ABSPATH . 'includes/wizards/class-dashboard.php';
 		include_once NEWSPACK_ABSPATH . 'includes/wizards/class-components-demo.php';
