+++ includes/wizards/class-setup-wizard.php
@@ -127,10 +127,10 @@ class Setup_Wizard extends Wizard {
 		);
 		register_rest_route(
 			NEWSPACK_API_NAMESPACE,
-			'/wizard/' . $this->slug . '/starter-content/categories',
+			'/wizard/' . $this->slug . '/starter-content/init',
 			[
 				'methods'             => WP_REST_Server::EDITABLE,
-				'callback'            => [ $this, 'api_starter_content_categories' ],
+				'callback'            => [ $this, 'api_starter_content_init' ],
 				'permission_callback' => [ $this, 'api_permissions_check' ],
 			]
 		);
@@ -158,24 +158,6 @@ class Setup_Wizard extends Wizard {
 			[
 				'methods'             => WP_REST_Server::EDITABLE,
 				'callback'            => [ $this, 'api_starter_content_homepage' ],
-				'permission_callback' => [ $this, 'api_permissions_check' ],
-			]
-		);
-		register_rest_route(
-			NEWSPACK_API_NAMESPACE,
-			'/wizard/' . $this->slug . '/existing-content/init',
-			[
-				'methods'             => WP_REST_Server::EDITABLE,
-				'callback'            => [ $this, 'api_existing_content_init' ],
-				'permission_callback' => [ $this, 'api_permissions_check' ],
-			]
-		);
-		register_rest_route(
-			NEWSPACK_API_NAMESPACE,
-			'/wizard/' . $this->slug . '/existing-content/post/(?P<id>[\a-z]+)',
-			[
-				'methods'             => WP_REST_Server::EDITABLE,
-				'callback'            => [ $this, 'api_existing_content_post' ],
 				'permission_callback' => [ $this, 'api_permissions_check' ],
 			]
 		);
@@ -245,13 +227,30 @@ class Setup_Wizard extends Wizard {
 	}
 
 	/**
-	 * Install starter content categories
+	 * Initialize a starter content generator.
 	 *
-	 * @return WP_REST_Response containing info.
+	 * @param WP_REST_Request $request API request object.
+	 * @return WP_REST_Response|WP_Error
 	 */
-	public function api_starter_content_categories() {
-		$status = Starter_Content::create_categories();
-		return rest_ensure_response( [ 'status' => $status ] );
+	public function api_starter_content_init( $request ) {
+		$request_params = $request->get_params();
+		if ( empty( $request_params ) || empty( $request_params['type'] ) ) {
+			return new WP_Error( 'newspack_setup', __( 'Missing starter content initialization info', 'newspack' ) );
+		}
+
+		$starter_content_type = 'import' === $request_params['type'] ? 'import' : 'generated';
+		$existing_site_url    = ! empty( $request_params['site'] ) && wp_http_validate_url( $request_params['site'] ) ? esc_url_raw( $request_params['site'] ) : '';
+
+		if ( 'import' === $starter_content_type && ! $existing_site_url ) {
+			return new WP_Error( 'newspack_setup', __( 'Invalid existing site URL.', 'newspack' ) );
+		}
+
+		$result = Starter_Content::initialize( $starter_content_type, $existing_site_url );
+		if ( is_wp_error( $result ) ) {
+			return $result;
+		}
+
+		return rest_ensure_response( [ 'status' => 200 ] );
 	}
 
 	/**
@@ -283,26 +282,6 @@ class Setup_Wizard extends Wizard {
 	 */
 	public function api_starter_content_homepage() {
 		$status = Starter_Content::create_homepage();
-		return rest_ensure_response( [ 'status' => $status ] );
-	}
-
-	public function api_existing_content_init( $request ) {
-		$request_params = $request->get_params();
-		if ( empty( $request_params ) || ! wp_http_validate_url( $request_params['site'] ) ) {
-			return new WP_Error( 'newspack_setup', __( 'Invalid site URL entered', 'newspack' ) );
-		}
-
-		$result = Starter_Content::initialize_existing_posts( $request_params['site'] );
-		if ( is_wp_error( $result ) ) {
-			return $result;
-		}
-
-		return rest_ensure_response( [ 'status' => 200 ] );
-	}
-
-	public function api_existing_content_post( $request ) {
-		$id = $request['id'];
-		$status = Starter_Content::create_existing_post( $id );
 		return rest_ensure_response( [ 'status' => $status ] );
 	}
